feat: 등록된 부스 배너 이미지 화면에 업데이트
