+++ apps/nowait-admin/src/pages/AdminBooth/AdminBooth.tsx
@@ -1,172 +1,12 @@
 import { useEffect, useState } from "react";
-import placeholderIcon from "../../assets/image_placeholder.svg";
 import { useWindowWidth } from "../../hooks/useWindowWidth";
-import OperatingTimeSelector from "./components/OperatingTimeSelector";
-import NoticeEditor from "./components/NoticeEditor";
 import MenuSection from "./components/MenuSection";
 import { useGetStore } from "../../hooks/booth/store/useGetStore";
 import { useUpdateStore } from "../../hooks/booth/store/useUpdateStore";
 import { useUploadStoreProfileImage } from "../../hooks/booth/store/useUploadStoreProfileImage";
 import { useUploadStoreBannerImages } from "../../hooks/booth/store/useUploadStoreBannerImages";
-import BoothProfileImage from "./components/BoothProfileImage";
-
-const BoothSection = ({
-  boothName,
-  setBoothName,
-  isFocused,
-  setIsFocused,
-  boothIntro,
-  setBoothIntro,
-  isTextareaFocused,
-  setIsTextareaFocused,
-  bannerImages,
-  setBannerImages,
-  boothNotice,
-  setBoothNotice,
-  startHour,
-  setStartHour,
-  startMinute,
-  setStartMinute,
-  endHour,
-  setEndHour,
-  endMinute,
-  setEndMinute,
-}: {
-  boothName: string;
-  setBoothName: (val: string) => void;
-  isFocused: boolean;
-  setIsFocused: (val: boolean) => void;
-  boothIntro: string;
-  setBoothIntro: (val: string) => void;
-  isTextareaFocused: boolean;
-  setIsTextareaFocused: (val: boolean) => void;
-  bannerImages: File[];
-  setBannerImages: (val: File[]) => void;
-  boothNotice: string;
-  setBoothNotice: (val: string) => void;
-  startHour: string;
-  setStartHour: (val: string) => void;
-  startMinute: string;
-  setStartMinute: (val: string) => void;
-  endHour: string;
-  setEndHour: (val: string) => void;
-  endMinute: string;
-  setEndMinute: (val: string) => void;
-}) => {
-  const [profileImage, setProfileImage] = useState<File | null>(null);
-  return (
-    <>
-      <div className="flex items-center py-[50px]">
-        <BoothProfileImage
-          profileImage={profileImage}
-          setProfileImage={setProfileImage}
-        />
-        <div className="flex flex-col w-[494px] ml-[50px] h-[115px]">
-          <span className="text-title-18-bold text-gray-500 mb-[6px] flex">
-            부스명
-          </span>
-          <span className="text-sm text-gray-500 mb-[14px] flex">
-            컴퓨터공학과
-          </span>
-          <div className="flex h-full relative">
-            <input
-              type="text"
-              value={boothName}
-              maxLength={20}
-              onFocus={() => setIsFocused(true)}
-              onBlur={() => setIsFocused(false)}
-              onChange={(e) => setBoothName(e.target.value)}
-              placeholder="부스명을 입력해주세요"
-              className="w-full h-full bg-black-5 border border-[#DDDDDD] rounded-xl px-4 py-2 text-sm bg-black-5 "
-            />
-            <span className="absolute right-[20px] top-1/2 -translate-y-1/2 text-xs text-gray-400">
-              <span className={isFocused ? "text-black" : "text-gray-400"}>
-                {boothName.length}
-              </span>{" "}
-              / 20
-            </span>
-          </div>
-        </div>
-      </div>
-
-      {/* 부스 소개 */}
-      <div className="flex flex-col mb-[50px] relative">
-        <label className="block font-semibold">부스 소개</label>
-        <p className="mt-[6px] mb-[14px] text-14-regular text-black-60">
-          부스를 자유롭게 소개해주세요
-        </p>
-        <textarea
-          maxLength={250}
-          className="w-full h-32 border border-[#DDDDDD] bg-black-5 bg-black-5 focus:bg-white rounded-lg text-sm px-[20px] pt-[16px] pr-[147px] pb-[33px]"
-          onFocus={() => setIsTextareaFocused(true)}
-          onBlur={() => setIsTextareaFocused(false)}
-          placeholder={isTextareaFocused ? "" : "부스 소개를 입력해주세요"}
-          value={boothIntro}
-          onChange={(e) => setBoothIntro(e.target.value)}
-        />
-        <div className="absolute bottom-[12px] right-[20px] text-right text-xs text-gray-400">
-          {boothIntro.length} / 250
-        </div>
-      </div>
-
-      {/* 배너 이미지 */}
-      <div className="flex flex-col mb-[50px]">
-        <label className="block font-semibold">배너 이미지</label>
-        <p className="text-14-regular text-black-60 mb-[14px]">
-          첫번째 이미지는 우리 부스를 대표하는 이미지로 설정돼요
-        </p>
-        <div className="flex gap-[10px]">
-          {Array(3)
-            .fill(null)
-            .map((_, i) => (
-              <label
-                key={i}
-                className="w-[150px] h-25 bg-black-5 border border-[#DDDDDD] rounded-xl flex items-center justify-center cursor-pointer"
-              >
-                <input
-                  type="file"
-                  className="hidden"
-                  onChange={(e) => {
-                    const file = e.target.files?.[0];
-                    if (file) {
-                      const newImages = [...bannerImages];
-                      newImages[i] = file;
-                      setBannerImages(newImages);
-                    }
-                  }}
-                />
-                {bannerImages[i] ? (
-                  <img
-                    src={URL.createObjectURL(bannerImages[i])}
-                    alt={`배너 ${i + 1}`}
-                    className="object-cover w-full h-full rounded-lg"
-                  />
-                ) : (
-                  <img src={placeholderIcon} className="" alt="업로드" />
-                )}
-              </label>
-            ))}
-        </div>
-      </div>
-
-      <OperatingTimeSelector
-        startHour={startHour}
-        setStartHour={setStartHour}
-        startMinute={startMinute}
-        setStartMinute={setStartMinute}
-        endHour={endHour}
-        setEndHour={setEndHour}
-        endMinute={endMinute}
-        setEndMinute={setEndMinute}
-      />
-
-      {/* 공지사항 */}
-      <NoticeEditor notice={boothNotice} setNotice={setBoothNotice} />
-
-      {/* 버튼 */}
-    </>
-  );
-};
+import type { BannerImage } from "./types/booth";
+import BoothSection from "./components/BoothSection";
 
 const BoothForm = () => {
   const width = useWindowWidth();
@@ -188,7 +28,7 @@ const BoothForm = () => {
 
   const [isFocused, setIsFocused] = useState(false);
   const [isTextareaFocused, setIsTextareaFocused] = useState(false);
-  const [bannerImages, setBannerImages] = useState<File[]>([]);
+  const [bannerImages, setBannerImages] = useState<BannerImage[]>([]);
 
   const handleSave = () => {
     updateStore(
@@ -202,15 +42,20 @@ const BoothForm = () => {
       },
       {
         onSuccess: () => {
-          // 프로필 이미지 업로드
-          if (bannerImages[0]) {
-            uploadProfileImage({ storeId, image: bannerImages[0] });
-          }
+          // 업로드할 파일만 필터링
+          const filesToUpload = bannerImages.filter(
+            (img): img is File => img instanceof File
+          );
 
-          // 나머지 배너 이미지 업로드
-          const bannerFiles = bannerImages.slice(1).filter(Boolean);
-          if (bannerFiles.length > 0) {
-            uploadBannerImages({ storeId, images: bannerFiles });
+          if (filesToUpload.length > 0) {
+            // 첫 번째 파일은 대표 프로필 이미지
+            uploadProfileImage({ storeId, image: filesToUpload[0] });
+
+            // 나머지는 배너 이미지
+            const restFiles = filesToUpload.slice(1);
+            if (restFiles.length > 0) {
+              uploadBannerImages({ storeId, images: restFiles });
+            }
           }
 
           alert("부스 정보가 성공적으로 저장되었습니다!");
@@ -225,10 +70,13 @@ const BoothForm = () => {
       setBoothName(store.name);
       setBoothIntro(store.description);
 
-      // 배너 이미지는 서버에서 URL로 받으므로 File 객체로 변환 불가 → 미리보기 용으로만 URL 보관
-      // 필요한 경우 bannerImages를 string[]으로 따로 관리
-      // 여기서는 단순히 console에 출력
-      console.log("배너 이미지:", store.bannerImages);
+      if (store.bannerImages) {
+        const formatted = store.bannerImages.map((img: any) => ({
+          ...img,
+          imageType: img.imageType ?? "BANNER", // 기본값
+        }));
+        setBannerImages(formatted);
+      }
     }
   }, [store]);
